Update - 20190630
1. update several functions;
2. fix bugs;

+++ WormProcessing/video/GetMapImage.m
@@ -2,7 +2,7 @@ function map_img = GetMapImage(need_map, src_imagename, map_range, map_filename)
 if need_map
 	src_image = imread(src_imagename);
     map_img = MapImageTo8Bit(src_image, map_range);
-    if ~isempty(map_filename)
+    if ~isempty(map_filename)% && ~exist(map_filename,'file')
         imwrite(map_img, map_filename);
     end
 else
